pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "setuptools_scm>=8"
]
build-backend = "setuptools.build_meta"

[project]
name = "programming-quiz-web-app"
authors = [
    {name = "Vincent Borden"},
    {name = "Lindsey Larvick"},
    {name = "Andrew Neugarten"}
]
description = "A web application for creating and taking programming quizzes."
readme = "README.md"
requires-python = ">=3.10"
keywords = ["quiz", "exam", "programming", "web"]
classifiers = [
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: Implementation :: CPython",
    "Framework :: Flask",
    "Topic :: Education :: Testing"
]
dependencies = [
    "Flask",
    "Flask-SQLAlchemy",
    "Flask-Migrate",
    "Flask-WTF[email]",
    "Flask-Talisman",
    "Flask-JWT-Extended",
    "SQLAlchemy",
    "mysql-connector-python",
    'bleach[css]',
    "PyJWT"
]
dynamic = ["version"]

[project.urls]
Repository = "https://github.com/larvickl/CS-467-Programming-Quiz.git"
Issues = "https://github.com/larvickl/CS-467-Programming-Quiz/issues"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_file = "src/programming_quiz_web_app/__version__.py"
